requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.0
audioread==3.0.1
babel==2.16.0
blinker==1.8.2
boto3==1.35.29
botocore==1.35.29
cached_path==1.6.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cn2an==0.5.22
contourpy==1.3.0
cycler==0.12.1
dateparser==1.1.8
decorator==5.1.1
Deprecated==1.2.14
Distance==0.1.3
docopt==0.6.2
eng-to-ipa==0.0.2
exceptiongroup==1.2.2
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.13.4
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.9.0
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gradio==4.44.0
gradio_client==1.3.0
grpcio==1.66.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.23.5
idna==3.10
importlib_resources==6.4.5
inflect==7.0.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.7
langid==1.1.6
librosa==0.9.1
llvmlite==0.43.0
loguru==0.7.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mecab-python3==1.0.9
melotts==0.1.2
mpmath==1.3.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plac==1.4.3
platformdirs==4.3.6
pooch==1.8.2
proces==0.1.7
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pykakasi==2.2.1
pyparsing==3.1.4
pypinyin==0.50.0
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
resampy==0.4.3
rich==13.8.1
rsa==4.9
ruff==0.6.8
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.38.6
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
transformers==4.27.4
triton==3.0.0
txtsplit==1.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.2.3
uvicorn==0.31.0
wasabi==0.10.1
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
